feat: make core IO agnostic
To be able to run PachaDB in several places (in-browser, deno,
cloudflare workers, rust projects, standalone, etc) it makes sense to
detach how it works from the underlying IO layer.

To do this I've introduced a few interfaces that help plug in different
parts of the IO layer as needed, and refactor the intake and query
workers to start using them.

I haven't tested this yet, but it at least compiles.

This also splits the Query stage into a Planner and an Executor, that we
can use for optimizing whatever plan we come up with in the future, and
for figuring out how to execute a query more efficiently (eg. parallely)
depending on the IO we're on.

As an example, on Cloudflare Workers we wouldn't be able to massively
parallelize by having a worker reach out to N workers, because there's
up to 6 concurrent fetch requests allowed.

However, locally, that would be fine.